moved createGroup to GroupView
