+++ public/js/service-common.js
@@ -21,33 +21,4 @@ export class CommonService {
       return "Unexpected response type";
     }
   }
-
-  /**
-   * Method used to create a group objest from values of the HTML elements
-   * @param {Element} groupHtmlElement The HTML content from which we want to create a group object
-   * @returns Object with group data retrieved from input HTML element
-   */
-  static createGroup(groupHtmlElement) {
-    const groupObservers = groupHtmlElement.querySelectorAll("div.modal-button:not(.new-observer)");
-    return {
-      name: groupHtmlElement.querySelector("input.group-name").value,
-      category: groupHtmlElement.querySelector("input.group-category").value,
-      domain: groupHtmlElement.querySelector("input.group-domain").value,
-      observers: this.#createGroupObservers(groupObservers),
-    };
-  }
-
-  /**
-   * Method used to create an array of observer objects from provided list of HTML elements
-   * @param {NodeList} observers The list of HTML observer elemenets
-   * @returns array of observer objects created from provided input
-   */
-  static #createGroupObservers(observers) {
-    const result = [];
-    observers.forEach((observer) => {
-      const observerContent = observer.parentNode.querySelector("div.modal-content");
-      result.push(ObserversView.fromHtml(observerContent));
-    });
-    return result;
-  }
 }
