+++ public/js/service-group.js
@@ -1,4 +1,5 @@
 import { CommonService } from "./service-common.js";
+import { GroupsView } from "./view-group.js";
 
 export class GroupsService {
   /**
@@ -77,7 +78,7 @@ export class GroupsService {
   static #createRequestOptions(method) {
     const shouldHaveBody = "POST" === method || "PUT" === method;
     const expandedGroup = document.querySelector("article.group-column.expanded");
-    const requestBody = shouldHaveBody ? JSON.stringify(CommonService.createGroup(expandedGroup)) : undefined;
+    const requestBody = shouldHaveBody ? JSON.stringify(GroupsView.createGroup(expandedGroup)) : undefined;
     return {
       method: method,
       headers: { "Content-Type": "application/json" },
